+++ game/src/game/obj_behaviors_2.c
@@ -5,6 +5,7 @@
 #include "actors/group11.h"
 #include "actors/group17.h"
 #include "audio/external.h"
+#include "audio/data.h"
 #include "behavior_actions.h"
 #include "behavior_data.h"
 #include "camera.h"